+++ src/Symfony/Component/Form/Extension/Validator/Type/RepeatedTypeValidatorExtension.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Form\Extension\Validator\Type;
 
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\OptionsResolver\Options;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
  * @author Bernhard Schussek <bschussek@gmail.com>
@@ -22,16 +23,16 @@ class RepeatedTypeValidatorExtension extends AbstractTypeExtension
     /**
      * {@inheritdoc}
      */
-    public function getDefaultOptions()
+    public function setDefaultOptions(OptionsResolver $resolver)
     {
         // Map errors to the first field
         $errorMapping = function (Options $options) {
             return array('.' => $options['first_name']);
         };
 
-        return array(
+        $resolver->setDefaults(array(
             'error_mapping' => $errorMapping,
-        );
+        ));
     }
 
     /**